requirements.txt
fastapi
uvicorn
pydantic
pyyaml
zowe-zos-files-for-zowe-sdk
zowe-apiml-onboarding-enabler-python